Feature: Export financial data as CSV
This was added as part of the calendar component in order to have the
button on the calendar's toolbar and because the calendar caches the
financial data. This allows us to avoid retrieving all the data again
from the backend API.

Similar to other aspects of the calendar component, this will need to be
refactored into a separate service in the future.